+++ extension/server/src/server.ts
@@ -17,7 +17,6 @@ import { URI } from 'vscode-uri';
 import completionItemProvider from './providers/completionItem';
 import hoverProvider from './providers/hover';
 import foldingRangeProvider from './providers/foldingRange';
-import stickyHeadersProvider from './providers/stickyHeaders';
 
 import { connection, getFileRequest, getObject as getObjectData, handleClientRequests, memberResolve, streamfileResolve, validateUri } from "./connection";
 import * as Linter from './providers/linter';
@@ -319,7 +318,6 @@ if (languageToolsEnabled) {
 	connection.onCodeAction(genericCodeActionsProvider);
 	connection.onSignatureHelp(signatureHelpProvider);
 	connection.onFoldingRanges(foldingRangeProvider);
-	connection.onDocumentSymbol(stickyHeadersProvider);
 
 	// project specific
 	connection.onWorkspaceSymbol(workspaceSymbolProvider);
